+++ src/main/java/cool/scx/_core/_cms/article/ArticleService.java
@@ -25,7 +25,7 @@ public class ArticleService extends BaseService<Article> {
     public Object testListAll(Map<String, Object> params) {
         var author = params.get("author") != null ? params.get("author").toString() : "";
         var p = new QueryParam();
-        Integer count = count(p);
+        Long count = count(p);
         if (count < 20) {
             var l = new ArrayList<Article>();
             for (int i = 0; i < 50; i++) {
